Resolving a circular dependency issue
